LayoutTests/media/modern-media-controls/macos-inline-media-controls/macos-inline-media-controls-buttons-styles-expected.txt

Testing the MacOSInlineMediaControls computed styles.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PlayPauseButton
PASS rgba(style.backgroundColor).r is 255
PASS rgba(style.backgroundColor).g is 255
PASS rgba(style.backgroundColor).b is 255
PASS rgba(style.backgroundColor).a is within 0.001 of 0.45
PASS style.height is "50px"
PASS style.webkitMaskPositionY is "12px"

SkipBackButton
PASS rgba(style.backgroundColor).r is 255
PASS rgba(style.backgroundColor).g is 255
PASS rgba(style.backgroundColor).b is 255
PASS rgba(style.backgroundColor).a is within 0.001 of 0.45
PASS style.height is "50px"
PASS style.webkitMaskPositionY is "10px"

MuteButton
PASS rgba(style.backgroundColor).r is 255
PASS rgba(style.backgroundColor).g is 255
PASS rgba(style.backgroundColor).b is 255
PASS rgba(style.backgroundColor).a is within 0.001 of 0.45
PASS style.height is "50px"
PASS style.webkitMaskPositionY is "10px"

AirplayButton
PASS rgba(style.backgroundColor).r is 255
PASS rgba(style.backgroundColor).g is 255
PASS rgba(style.backgroundColor).b is 255
PASS rgba(style.backgroundColor).a is within 0.001 of 0.45
PASS style.height is "50px"
PASS style.webkitMaskPositionY is "13px"

PiPButton
PASS rgba(style.backgroundColor).r is 255
PASS rgba(style.backgroundColor).g is 255
PASS rgba(style.backgroundColor).b is 255
PASS rgba(style.backgroundColor).a is within 0.001 of 0.45
PASS style.height is "50px"
PASS style.webkitMaskPositionY is "13px"

TracksButton
PASS rgba(style.backgroundColor).r is 255
PASS rgba(style.backgroundColor).g is 255
PASS rgba(style.backgroundColor).b is 255
PASS rgba(style.backgroundColor).a is within 0.001 of 0.45
PASS style.height is "50px"
PASS style.webkitMaskPositionY is "15px"

FullscreenButton
PASS rgba(style.backgroundColor).r is 255
PASS rgba(style.backgroundColor).g is 255
PASS rgba(style.backgroundColor).b is 255
PASS rgba(style.backgroundColor).a is within 0.001 of 0.45
PASS style.height is "50px"
PASS style.webkitMaskPositionY is "13px"

PASS successfullyParsed is true

TEST COMPLETE
